requirements.txt
openai>=1.0.0
pinecone-client>=5.0.0
requests>=2.31.0
